refactor: EventFeedParser.cs almost done, add a Loader window to load resolver config file (on going)
